services/vuln-detector/requirements.txt
fastapi
uvicorn
python-dotenv
pyyaml
psycopg2-binary
